+++ dfm_tools/xugrid_helpers.py
@@ -260,6 +260,7 @@ def open_partitioned_dataset(file_nc, decode_fillvals=False, remove_edges=True, 
 
 def open_dataset_curvilinear(file_nc,
                              varn_lon='longitude',
+                             varn_lat='latitude',
                              varn_vert_lon='vertices_longitude', #'grid_x'
                              varn_vert_lat='vertices_latitude', #'grid_y'
                              ij_dims=['i','j'], #['N','M']
@@ -269,11 +270,13 @@ def open_dataset_curvilinear(file_nc,
     This is a first version of a function that creates a xugrid UgridDataset from a curvilinear dataset like CMCC. Curvilinear means in this case 2D lat/lon variables and i/j indexing. The CMCC dataset does contain vertices, which is essential for conversion to ugrid.
     It also works for WAQUA files that are converted with getdata
     """
+    # TODO: maybe get varn_lon/varn_lat automatically with cf-xarray (https://github.com/xarray-contrib/cf-xarray)
     
     if 'chunks' not in kwargs:
         kwargs['chunks'] = {'time':1}
     
-    ds = xr.open_mfdataset(file_nc, **kwargs)
+    # data_vars='minimal' to avoid time dimension on vertices_latitude and others when opening multiple files at once
+    ds = xr.open_mfdataset(file_nc, data_vars="minimal", **kwargs)
     
     print('>> getting vertices from ds: ',end='')
     dtstart = dt.datetime.now()
@@ -285,8 +288,7 @@ def open_dataset_curvilinear(file_nc,
     
     # convert from 0to360 to -180 to 180
     if convert_360to180:
-        vertices_longitude = (vertices_longitude+180)%360 - 180
-        ds[varn_lon] = (ds[varn_lon] + 180) % 360 - 180
+        vertices_longitude = (vertices_longitude+180) % 360 - 180
     
     # face_xy = np.stack([longitude,latitude],axis=-1)
     # face_coords_x, face_coords_y = face_xy.T
@@ -323,14 +325,22 @@ def open_dataset_curvilinear(file_nc,
     print('>> stacking ds i/j coordinates: ',end='') #fast
     dtstart = dt.datetime.now()
     face_dim = grid.face_dimension
-    ds_stacked = ds.stack({face_dim:ij_dims}).sel({face_dim:bool_combined}) #TODO: lev/time bnds are dropped, avoid this. maybe stack initial dataset since it would also simplify the rest of the function a bit
-    ds_stacked = ds_stacked.drop_vars(ij_dims+['mesh2d_nFaces']) #TODO: solve "DeprecationWarning: Deleting a single level of a MultiIndex is deprecated", solved by removing mesh2d_nFaces variable
+    # TODO: lev/time bnds are dropped, avoid this. maybe stack initial dataset since it would also simplify the rest of the function a bit
+    ds_stacked = ds.stack({face_dim:ij_dims}).sel({face_dim:bool_combined})
+    latlon_vars = [varn_lon, varn_lat, varn_vert_lon, varn_vert_lat]
+    ds_stacked = ds_stacked.drop_vars(ij_dims + latlon_vars)
     print(f'{(dt.datetime.now()-dtstart).total_seconds():.2f} sec')
     
     print('>> init uds: ',end='') #long
     dtstart = dt.datetime.now()
     uds = xu.UgridDataset(ds_stacked,grids=[grid])
     print(f'{(dt.datetime.now()-dtstart).total_seconds():.2f} sec')
+    
+    # drop 0-area cells (relevant for CMCC global datasets)
+    bool_zero_cell_size = uds.grid.area==0
+    if bool_zero_cell_size.any():
+        print(f"WARNING: dropping {bool_zero_cell_size.sum()} 0-sized cells from dataset")
+        uds = uds.isel({uds.grid.face_dimension: ~bool_zero_cell_size})
     
     #remove faces that link to node coordinates that are nan (occurs in waqua models)
     bool_faces_wnannodes = np.isnan(uds.grid.face_node_coordinates[:,:,0]).any(axis=1)